requirements.txt
django
djangorestframework
numpy
requests
pillow
torch
torchvision
